drvEIDE: use shared hd registry

+++ src/drivers-i386/ide/drvEIDE/EIDE.drvproj/PostLoad.tproj/PostLoad.m
@@ -25,9 +25,8 @@
  * Copyright 1997-1998 by Apple Computer, Inc., All rights reserved.
  * Copyright 1994-1997 NeXT Software, Inc., All rights reserved.
  *
- * Post-Load command for IDE disks.  Creates device nodes /dev/hd[0-3][a-h]
- * and /dev/rhd[0-3][a-h].
- * units.
+ * Post-Load command for ATA disks. Creates device nodes /dev/hd[0-31][a-h]
+ * and /dev/rhd[0-31][a-h].
  *
  * HISTORY
  * 30-Sep-94    Rakesh Dubey at NeXT
@@ -35,14 +34,11 @@
  */
 
 #import <streams/streams.h>
-#import <driverkit/IODeviceMaster.h>
-#import <driverkit/IODevice.h>
 #import <errno.h>
 #import <libc.h>
+#import <sys/stat.h>
 
-#import <ata_extern.h>
-
-#define PATH_NAME_SIZE 			10
+#define PATH_NAME_SIZE 			16
 #define DEV_STRING 			"/dev/"
 #define IDE_INIT_ERR_STRING 		"Error initializing IDE driver"
 
@@ -53,84 +49,28 @@
 
 #define NIDE_PARTITIONS 		8
 
-#define NIDE_DEVICES			(MAX_IDE_DRIVES * MAX_IDE_CONTROLLERS)
+#define NIDE_DEVICES			32
+
+#define IDE_BLOCK_MAJOR			3
+#define IDE_CHARACTER_MAJOR		15
 
 static int makeNode(char *deviceName, int iUnit, int major, int num,
 			unsigned short mode);
 
 int main(int argc, char **argv)
 {
-    IOReturn			ret;
-    IOObjectNumber		tag;
-    IOString			kind;
-    unsigned int		count;
-    IODeviceMaster		*devMaster;
-    int				blockMajor, characterMajor;
     int				i, iUnit, iRet;
-    char 			path[PATH_NAME_SIZE];
 
     iRet = 0;
-    
-    devMaster = [IODeviceMaster new];
 
-    /*
-     * The kernel creates first two. This needs some investigation. FIXME.
-     */
-    for (iUnit = 2; iUnit < NIDE_DEVICES; iUnit ++) {
-
-	bzero(path, PATH_NAME_SIZE);
-	sprintf(path, "%s%s%d", DEV_STRING, "hd", iUnit);
-
-	/*
-	 * Find this instance of the IDE driver
-	 */
-
-	ret = [devMaster lookUpByDeviceName: path + strlen(DEV_STRING)
-	    objectNumber: &tag 
-	    deviceKind: &kind];
-
-#ifdef DEBUG
-printf ("unit %d, obj name %s\n", iUnit, path);
-printf ("ret = %d\n", ret);
-#endif DEBUG
-
-	/*
-	 * If this device does not exist then we don't want to create nodes
-	 * for it. This saves some time.
-	 */
-	if (ret != IO_R_SUCCESS) {
-	    continue;
-	}
-	
-	/*
-	 * Query the object for its major device number 
-	 */
-	blockMajor = -1;
-	ret = [devMaster getIntValues:&blockMajor
-	       forParameter:"BlockMajor" objectNumber:tag
-	       count:&count];
-	if (ret != IO_R_SUCCESS) {
-	    printf("%s: couldn't get block major number:  Returned %d.\n",
-		   IDE_INIT_ERR_STRING, ret);
-	    iRet = -1;
-	}
-	
-	characterMajor = -1;
-	ret = [devMaster getIntValues:&characterMajor
-	       forParameter:"CharacterMajor" objectNumber:tag
-	       count:&count];
-	if (ret != IO_R_SUCCESS) {
-	    printf("%s: couldn't get char major number:  Returned %d.\n",
-		   IDE_INIT_ERR_STRING, ret);
-	    iRet = -1;
-	}
-	
-	/*
-	 * Remove old devs and create new ones for this unit. 	 
-	 */
+    for (iUnit = 0; iUnit < NIDE_DEVICES; iUnit ++) {
 	for (i = 0; i < NIDE_PARTITIONS; i++) {
-	    iRet = makeNode("hd", iUnit, blockMajor, i, DEV_MOD_BLOCK);
-	    iRet = makeNode("rhd", iUnit, characterMajor, i, DEV_MOD_CHAR);
+	    if (makeNode("hd", iUnit, IDE_BLOCK_MAJOR, i,
+			 DEV_MOD_BLOCK) != 0)
+		iRet = -1;
+	    if (makeNode("rhd", iUnit, IDE_CHARACTER_MAJOR, i,
+			 DEV_MOD_CHAR) != 0)
+		iRet = -1;
 	}
     }
     
@@ -140,11 +80,10 @@ printf ("ret = %d\n", ret);
 static int makeNode(char *deviceName, int iUnit, int major, int num,
 			unsigned short mode)
 {
-    int	iRet;
+    struct stat status;
+    dev_t device;
     int minor;
     char path[PATH_NAME_SIZE];
-    
-    iRet = -1;
     
 #ifdef notdef
 printf ("makeNode %s, iUnit %d num %d\n", deviceName, iUnit, num);
@@ -154,21 +93,31 @@ printf ("makeNode %s, iUnit %d num %d\n", deviceName, iUnit, num);
     sprintf(path, "%s%s%d%c", DEV_STRING, deviceName, iUnit, num + 'a');
 
     minor = iUnit * NIDE_PARTITIONS + num;
-    
-    if (unlink(path)) {
-	if (errno != ENOENT) {
+    device = (major << 8) | minor;
+
+    if (lstat(path, &status) == 0) {
+	if ((status.st_mode & S_IFMT) == (mode & S_IFMT) &&
+	    (status.st_mode & 07777) == (mode & 07777) &&
+	    status.st_rdev == device)
+	    return 0;
+	if (unlink(path)) {
 	    printf("%s: could not delete old %s.  Errno is %d\n",
-		IDE_INIT_ERR_STRING, path, errno);
-	    iRet = -1; 
+		   IDE_INIT_ERR_STRING, path, errno);
+	    return -1;
 	}
+
+    } else if (errno != ENOENT) {
+	printf("%s: could not inspect %s.  Errno is %d\n",
+	       IDE_INIT_ERR_STRING, path, errno);
+	return -1;
     }
 
     umask(DEV_UMASK);
-    if (mknod(path, mode, (major << 8) | minor)) {
+    if (mknod(path, mode, device)) {
 	printf("%s: could not create %s.  Errno is %d\n",
 	    IDE_INIT_ERR_STRING, path, errno);
-	iRet = -1;
+	return -1;
     }
-    
-    return iRet;
+
+    return 0;
 }
